# [GPU] Add b_fs_yx_fsv16 fp16 onednn concatenation testcase