### Details:
 - onednn concatenation failed with some shapes when using b_fs_yx_fsv16 layout, like 1x88x52x52.

### Tickets:
 - CVS-169075