@wilson-seok 2025-08-19
0 writing in eltwise kernel should impact to all eltwise blocked_opt/ref kernel execution time when blocked format. Please check kernel time for several input shapes(small to large).