@yuanxion 2025-08-22
> 1. 0 writing in eltwise kernel should impact to all eltwise blocked_opt/ref kernel execution time when blocked format. Please check kernel time for several input shapes(small to large).
> 2. Could you please update PR description to include how the non-zero value in blocked format padded area makes accuracy issue in concat. You can add onednn primitive description also.

@wilson-seok Updated the descriptions.
The performance test shown that the performance degradation for eltwise_blocked_opt is not big, but for generic_eltwise_ref it is very large.

new kernel performance (%) | iters 1 | iters 10 | iters 20 | iters 30
-- | -- | -- | -- | --
eltwise_blocked_opt   1x176x52x52_crop_b_fs_yx_fsv16 | 99.2% | 98.3% | 97.5% | 98.7%
eltwise_blocked_opt   1x352x114x114_crop_b_fs_yx_fsv16 | 99.6% | 99.1% | 94.8% | 100.3%
eltwise_blocked_opt   1x200x114x114_crop_b_fs_yx_fsv16 | 96.2% | 97.7% | 99.1% | 97.5%
eltwise_blocked_opt   1x200x512x512_crop_b_fs_yx_fsv16 | 103.6% | 102.0% | 101.0% | 103.8%
eltwise_blocked_opt   1x200x1024x1024_crop_b_fs_yx_fsv16 | 105.7% | 100.5% | 99.8% | 99.4%
generic_eltwise_ref     1x200x114x114_crop_b_fs_yx_fsv16 | 253.3% | 289.9% | 270.3% | 167.7%
generic_eltwise_ref     1x200x512x512_crop_b_fs_yx_fsv16 | 287.2% | 295.0% | 288.8% | 290.9%
generic_eltwise_ref   1x200x1024x1024_crop_b_fs_yx_fsv16 | 110.4% | 110.4% | 110.6% | 110.7%