@yuanxion 2025-08-25
Further optimized the performance for **generic_eltwise_ref** kernel using multiple threads for zero-padding, and its performance degradation (eltwise kernel itself, not E2E test) dropped from about 250% ~ 290% (difference shapes) to about 150% ~ 100%.

new kernel performance (%) | iters 1 | iters 10 | iters 20 | iters 30
-- | -- | -- | -- | --
generic_eltwise_ref   1x200x52x52_crop_b_fs_yx_fsv16 | 148.7% | 140.4% | 140.5% | 139.7%
generic_eltwise_ref   1x200x114x114_crop_b_fs_yx_fsv16 | 136.9% | 134.4% | 129.7% | 128.0%
generic_eltwise_ref   1x200x512x512_crop_b_fs_yx_fsv16 | 98.9% | 99.1% | 106.2% | 106.1%
generic_eltwise_ref  1x200x1024x1024_crop_b_fs_yx_fsv16 | 98.8% | 99.4% | 98.7% | 99.3%
eltwise_blocked_opt   1x176x52x52_crop_b_fs_yx_fsv16 | 111.0% | 109.6% | 111.4% | 110.2%
eltwise_blocked_opt   1x176x52x52_crop_b_fs_yx_fsv16 | 99.7% | 98.5% | 98.3% | 97.6%
eltwise_blocked_opt   1x200x52x52_crop_b_fs_yx_fsv16 | 99.0% | 99.7% | 103.0% | 102.3%
eltwise_blocked_opt   1x200x114x114_crop_b_fs_yx_fsv16 | 99.0% | 98.9% | 99.9% | 100.7%
eltwise_blocked_opt   1x200x512x512_crop_b_fs_yx_fsv16 | 100.5% | 99.0% | 102.3% | 101.0%
eltwise_blocked_opt   1x200x1024x1024_crop_b_fs_yx_fsv16 | 105.5% | 101.4% | 100.1% | 100.6%