@yuanxion 2025-08-29
The performance test result for models with benchmark_app shows slightly performance drop with this PR:

Throughput/FPS | original | after-fix | ratio
-- | -- | -- | --
hbonet0.5-FP16 | 293.674 | 290.044 | 98.8%
hbonet1.0-FP16 | 303.176 | 298.59 | 98.5%
hbonet1.0-FP16-INT8 | 237.994 | 235.228 | 98.8%
nanodet | 259.588 | 255.878 | 98.6%